fix: remove creationMessage in case we initiated the event

+++ src/script/conversation/ConversationRepository.ts
@@ -109,7 +109,7 @@ import {User} from '../entity/User';
 import {EventService} from '../event/EventService';
 import {ConnectionEntity} from '../connection/ConnectionEntity';
 import {EventSource} from '../event/EventSource';
-import {MemberMessage} from '../entity/message/MemberMessage';
+import {isMemberMessage, MemberMessage} from '../entity/message/MemberMessage';
 import {FileAsset} from '../entity/message/FileAsset';
 import type {EventRecord} from '../storage';
 import {MessageRepository} from './MessageRepository';
@@ -618,25 +618,38 @@ export class ConversationRepository {
 
     const mappedMessageEntities = await this.addEventsToConversation(events, conversationEntity);
     conversationEntity.hasAdditionalMessages(hasAdditionalMessages);
+
     if (!hasAdditionalMessages) {
-      const firstMessage = conversationEntity.getFirstMessage() as MemberMessage;
-      const checkCreationMessage = firstMessage?.isMember() && firstMessage.isCreation();
-      if (checkCreationMessage) {
-        const groupCreationMessageIn1to1 = conversationEntity.is1to1() && firstMessage.isGroupCreation();
-        const one2oneConnectionMessageInGroup = conversationEntity.isGroup() && firstMessage.isConnection();
-        const wrongMessageTypeForConversation = groupCreationMessageIn1to1 || one2oneConnectionMessageInGroup;
+      const allMessages = conversationEntity.getAllMessages();
+      if (!!allMessages.length) {
+        const firstMessage = allMessages[0];
+        if (isMemberMessage(firstMessage)) {
+          console.info('adrian', conversationEntity, firstMessage, allMessages);
+          const checkCreationMessage = firstMessage.isMember() && firstMessage.isCreation();
+          if (checkCreationMessage) {
+            const groupCreationMessageIn1to1 = conversationEntity.is1to1() && firstMessage.isGroupCreation();
 
-        if (wrongMessageTypeForConversation) {
-          this.messageRepository.deleteMessage(conversationEntity, firstMessage);
-          conversationEntity.hasCreationMessage = false;
-        } else {
-          conversationEntity.hasCreationMessage = true;
+            const one2oneConnectionMessageInGroup = conversationEntity.isGroup() && firstMessage.isConnection();
+            const wrongMessageTypeForConversation = groupCreationMessageIn1to1 || one2oneConnectionMessageInGroup;
+
+            if (wrongMessageTypeForConversation) {
+              this.messageRepository.deleteMessage(conversationEntity, firstMessage);
+              conversationEntity.hasCreationMessage = false;
+            } else {
+              conversationEntity.hasCreationMessage = true;
+            }
+          }
+
+          // We know that we injected the creation message in this case, so we can remove it from the list
+          if (conversationEntity.isCreatedBySelf()) {
+            conversationEntity.hasCreationMessage = true;
+          }
+
+          if (!conversationEntity.hasCreationMessage) {
+            console.info('adrian', 'add message', conversationEntity.hasCreationMessage);
+            this.addCreationMessage(conversationEntity, this.userState.self().isTemporaryGuest());
+          }
         }
-      }
-
-      const addCreationMessage = !conversationEntity.hasCreationMessage;
-      if (addCreationMessage) {
-        this.addCreationMessage(conversationEntity, this.userState.self().isTemporaryGuest());
       }
     }
     return mappedMessageEntities;
